fix 'cd' command
