+++ .gitignore
@@ -14,3 +14,4 @@ a.out
 moshell
 *.bin
 completions
+sandbox
